Use bunyan for optional logging. (#19)
The default settings in bin/user-agent-service are for local developers,
so they log aggressively to stdout.  There's a corresponding Tofino
patch to pipe the UAS process stdout to the Tofino main process stdout,
and to turn the amount of logging down in that situation.

I elected to make the logger optional rather than get it working in
core.cljs for now; it's not hard, but I'll do it when I next want to
develop using figwheel.